Refactor files into pages, components, and data incl(fetchers, files, and logging)
