+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './data/reportWebVitals';
 
 import { Amplify } from 'aws-amplify';
 import amplifyconfig from './amplifyconfiguration.json';
